requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
pydantic
pytest
python-dotenv
